final_project/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
python-dotenv==0.13.0
pyJWT==1.7.1
